refactor: refactors integration tests
This create a separate file with integration tests helpers that are
shared between validate cluster tests and generate 1-1-restore tests.
