+++ pkg/service/one2onerestore/service_integration_test.go
@@ -6,28 +6,17 @@
 package one2onerestore
 
 import (
-	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"testing"
-	"time"
 
 	"github.com/pkg/errors"
-	"github.com/scylladb/go-log"
 	"github.com/scylladb/gocqlx/v2"
 	"github.com/scylladb/gocqlx/v2/qb"
-	"github.com/scylladb/scylla-manager/v3/pkg/metrics"
-	"github.com/scylladb/scylla-manager/v3/pkg/scyllaclient"
-	"github.com/scylladb/scylla-manager/v3/pkg/service/backup"
 	. "github.com/scylladb/scylla-manager/v3/pkg/service/backup/backupspec"
-	"github.com/scylladb/scylla-manager/v3/pkg/service/cluster"
 	. "github.com/scylladb/scylla-manager/v3/pkg/testutils"
 	. "github.com/scylladb/scylla-manager/v3/pkg/testutils/db"
 	. "github.com/scylladb/scylla-manager/v3/pkg/testutils/testconfig"
-	. "github.com/scylladb/scylla-manager/v3/pkg/testutils/testhelper"
-	"github.com/scylladb/scylla-manager/v3/pkg/util/uuid"
-	"go.uber.org/zap/zapcore"
 )
 
 func TestOne2OneRestoreServiceIntegration(t *testing.T) {
@@ -64,7 +53,7 @@ func TestOne2OneRestoreServiceIntegration(t *testing.T) {
 		"location":          loc,
 		"snapshot_tag":      tag,
 		"source_cluster_id": h.clusterID,
-		"nodes_mapping":     h.getNodeMappings(t),
+		"nodes_mapping":     getNodeMappings(t, h.client),
 	})
 
 	Print("Validate data")
@@ -72,161 +61,6 @@ func TestOne2OneRestoreServiceIntegration(t *testing.T) {
 	if srcCnt != dstCnt {
 		t.Fatalf("Expected row count %d, but got %d", srcCnt, dstCnt)
 	}
-}
-
-func testLocation(bucket, dc string) Location {
-	return Location{
-		DC:       dc,
-		Provider: S3,
-		Path:     "restoretest-" + bucket,
-	}
-}
-
-func newBackupSvc(t *testing.T, mgrSession gocqlx.Session, client *scyllaclient.Client, clusterID uuid.UUID) *backup.Service {
-	svc, err := backup.NewService(
-		mgrSession,
-		defaultBackupTestConfig(),
-		metrics.NewBackupMetrics(),
-		func(_ context.Context, id uuid.UUID) (string, error) {
-			return "test_cluster", nil
-		},
-		func(context.Context, uuid.UUID) (*scyllaclient.Client, error) {
-			return client, nil
-		},
-		func(ctx context.Context, clusterID uuid.UUID, _ ...cluster.SessionConfigOption) (gocqlx.Session, error) {
-			return CreateSession(t, client), nil
-		},
-		NewTestConfigCacheSvc(t, clusterID, client.Config().Hosts),
-		log.NewDevelopmentWithLevel(zapcore.InfoLevel).Named("backup"),
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return svc
-}
-
-func defaultBackupTestConfig() backup.Config {
-	return backup.Config{
-		DiskSpaceFreeMinPercent:   5,
-		LongPollingTimeoutSeconds: 1,
-		AgeMax:                    24 * time.Hour,
-	}
-}
-
-func newRestoreSvc(t *testing.T, mgrSession gocqlx.Session, client *scyllaclient.Client, clusterID uuid.UUID, user, pass string) *Service {
-	configCacheSvc := NewTestConfigCacheSvc(t, clusterID, client.Config().Hosts)
-
-	svc, err := NewService(
-		mgrSession,
-		func(context.Context, uuid.UUID) (*scyllaclient.Client, error) {
-			return client, nil
-		},
-		func(ctx context.Context, clusterID uuid.UUID, _ ...cluster.SessionConfigOption) (gocqlx.Session, error) {
-			return CreateManagedClusterSession(t, false, client, user, pass), nil
-		},
-		configCacheSvc,
-		log.NewDevelopmentWithLevel(zapcore.InfoLevel).Named("1-1-restore"),
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return svc
-}
-
-func newTestHelper(t *testing.T, hosts []string) *testHelper {
-	clientCfg := scyllaclient.TestConfig(hosts, AgentAuthToken())
-	sc, err := scyllaclient.NewClient(clientCfg, log.NopLogger)
-	requireNoError(t, err)
-	session := CreateScyllaManagerDBSession(t)
-
-	clusterID := uuid.NewTime()
-
-	backupSvc := newBackupSvc(t, session, sc, clusterID)
-	restoreSvc := newRestoreSvc(t, session, sc, clusterID, "", "")
-
-	return &testHelper{
-		client:     sc,
-		clusterID:  clusterID,
-		backupSvc:  backupSvc,
-		restoreSvc: restoreSvc,
-	}
-}
-
-type testHelper struct {
-	client     *scyllaclient.Client
-	clusterID  uuid.UUID
-	backupSvc  *backup.Service
-	restoreSvc *Service
-}
-
-func (h *testHelper) runBackup(t *testing.T, props map[string]any) string {
-	t.Helper()
-	Printf("Run backup with properties: %v", props)
-	ctx := context.Background()
-	taskID := uuid.NewTime()
-	runID := uuid.NewTime()
-
-	rawProps, err := json.Marshal(props)
-	if err != nil {
-		t.Fatal(errors.Wrap(err, "marshal properties"))
-	}
-
-	target, err := h.backupSvc.GetTarget(ctx, h.clusterID, rawProps)
-	if err != nil {
-		t.Fatal(errors.Wrap(err, "generate target"))
-	}
-
-	err = h.backupSvc.Backup(ctx, h.clusterID, taskID, runID, target)
-	if err != nil {
-		t.Fatal(errors.Wrap(err, "run backup"))
-	}
-
-	pr, err := h.backupSvc.GetProgress(ctx, h.clusterID, taskID, runID)
-	if err != nil {
-		t.Fatal(errors.Wrap(err, "get progress"))
-	}
-
-	return pr.SnapshotTag
-}
-
-func (h *testHelper) runRestore(t *testing.T, props map[string]any) {
-	t.Helper()
-	Printf("Run 1-1-restore with properties: %v", props)
-	ctx := context.Background()
-	taskID := uuid.NewTime()
-	runID := uuid.NewTime()
-
-	rawProps, err := json.Marshal(props)
-	if err != nil {
-		t.Fatal(errors.Wrap(err, "marshal properties"))
-	}
-
-	err = h.restoreSvc.One2OneRestore(ctx, h.clusterID, taskID, runID, rawProps)
-	if err != nil {
-		t.Fatal(errors.Wrap(err, "run 1-1-restore"))
-	}
-}
-
-func (h *testHelper) getNodeMappings(t *testing.T) []nodeMapping {
-	t.Helper()
-	ctx := context.Background()
-
-	var result []nodeMapping
-
-	nodesStatus, err := h.client.Status(ctx)
-	requireNoError(t, err)
-
-	for _, n := range nodesStatus {
-		rack, err := h.client.HostRack(ctx, n.Addr)
-		requireNoError(t, err)
-		result = append(result, nodeMapping{
-			Source: node{DC: n.Datacenter, Rack: rack, HostID: n.HostID},
-			Target: node{DC: n.Datacenter, Rack: rack, HostID: n.HostID},
-		})
-	}
-
-	return result
 }
 
 func truncateAllTablesInKeyspace(tb testing.TB, session gocqlx.Session, ks string) {
